implement parser for settings section
